.history/airflow_project/requirements_20251005170126.txt
apache-airflow==2.8.1
pandas
psycopg2-binary
matplotlib
seaborn
numpy
